Added Manage

+++ rail_frontend/src/components/AppointmentInfo.js
@@ -10,11 +10,14 @@ const AppointmentInfo = ({ appointment, onDeleteAppointment }) => {
     let delete_button;
     let book_now;
     if (Cookies.get('username')){
+      if (Cookies.get('is_admin')){
         delete_button = <button onClick={() => onDeleteAppointment(appointment.id)} type="button"
         className="p-1.5 mr-1.5 mt-1 rounded text-white bg-red-500 hover:bg-yellow-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500">
         <BiTrash /></button>;
 
-        book_now = <button class="px-4 py-1 text-sm text-purple-600 font-semibold rounded-full border border-purple-200 hover:text-white hover:bg-purple-600 hover:border-transparent focus:outline-none focus:ring-2 focus:ring-purple-600 focus:ring-offset-2"><Link to="/bookTrain">Book Now</Link></button>;
+      }
+
+      book_now = <Link to="/bookNow"><button class="px-4 py-1 text-sm text-purple-600 font-semibold rounded-full border border-purple-200 hover:text-white hover:bg-purple-600 hover:border-transparent focus:outline-none focus:ring-2 focus:ring-purple-600 focus:ring-offset-2">Book Now</button></Link>;
         
     
     }else{
